requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
clickclick==1.2.2
connexion==2.6.0
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
docutils==0.15.2
Flask==1.1.1
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
importlib-metadata==1.5.0
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.0
jsonschema==3.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.17.3
oauthlib==3.1.0
openapi-spec-validator==0.2.8
opt-einsum==3.1.0
pandas==0.25.2
patsy==0.5.1
Pillow==6.2.1
pmdarima==1.5.2
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.2
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
sklearn==0.0
statistics==1.0.3.5
statsmodels==0.10.2
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==3.1.0
